requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==0.16
Jinja2==2.7
MarkupSafe==0.18
SQLAlchemy==0.8.2
Unidecode==0.04.14
Werkzeug==0.9.2
gunicorn==17.5
itsdangerous==0.22
mongokit==0.9.0
pymongo==2.6.3
python-dateutil==2.1
python-slugify==0.0.6
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
